+++ compiler/typecheck/TcHsSyn.hs
@@ -1264,8 +1264,6 @@ zonkEvTerm env (EvTypeable ty ev) =
 zonkEvTerm env (EvCallStack cs)
   = case cs of
       EvCsEmpty -> return (EvCallStack cs)
-      EvCsTop n l tm -> do { tm' <- zonkEvTerm env tm
-                           ; return (EvCallStack (EvCsTop n l tm')) }
       EvCsPushCall n l tm -> do { tm' <- zonkEvTerm env tm
                                 ; return (EvCallStack (EvCsPushCall n l tm')) }