+++ libraries/base/GHC/IO/Exception.hs
@@ -357,7 +357,7 @@ assertError predicate v
   | predicate = lazy v
   | otherwise = throw (AssertionFailed
                         ("Assertion failed\n"
-                         ++ showCallStack (popCallStack ?callStack)))
+                         ++ prettyCallStack ?callStack))
 
 unsupportedOperation :: IOError
 unsupportedOperation =