+++ libraries/base/GHC/Stack/Types.hs
@@ -18,7 +18,9 @@
 
 module GHC.Stack.Types (
     -- * Implicit parameter call stacks
-    SrcLoc(..), CallStack(..),
+    CallStack, getCallStack, pushCallStack,
+    -- * Source locations
+    SrcLoc(..)
   ) where
 
 {-
@@ -44,22 +46,28 @@ import GHC.Integer ()
 -- Explicit call-stacks built via ImplicitParams
 ----------------------------------------------------------------------
 
--- | @CallStack@s are an alternate method of obtaining the call stack at a given
--- point in the program.
+-- | Implicit @CallStack@s are an alternate method of obtaining the call stack
+-- at a given point in the program.
 --
--- When an implicit-parameter of type @CallStack@ occurs in a program, GHC will
--- solve it with the current location. If another @CallStack@ implicit-parameter
--- is in-scope (e.g. as a function argument), the new location will be appended
--- to the one in-scope, creating an explicit call-stack. For example,
+-- GHC has two built-in rules for solving implicit-parameters of type
+-- @CallStack@.
+--
+-- 1. If the @CallStack@ occurs in a function call, it appends the
+--    source location of the call to the @CallStack@ in the environment.
+-- 2. @CallStack@s that cannot be solved normally (i.e. unbound
+--    occurrences) are defaulted to the empty @CallStack@.
+--
+-- Otherwise implicit @CallStack@s behave just like ordinary implicit
+-- parameters. For example:
 --
 -- @
--- myerror :: (?loc :: CallStack) => String -> a
--- myerror msg = error (msg ++ "\n" ++ showCallStack ?loc)
+-- myerror :: (?callStack :: CallStack) => String -> a
+-- myerror msg = error (msg ++ "\n" ++ prettyCallStack ?callStack)
 -- @
+--
 -- ghci> myerror "die"
 -- *** Exception: die
--- CallStack:
---   ?loc, called at MyError.hs:7:51 in main:MyError
+-- CallStack (from ImplicitParams):
 --   myerror, called at <interactive>:2:1 in interactive:Ghci1
 --
 -- @CallStack@s do not interact with the RTS and do not require compilation with
@@ -71,13 +79,38 @@ import GHC.Integer ()
 -- function that was called, the 'SrcLoc' is the call-site. The list is
 -- ordered with the most recently called function at the head.
 --
--- @since 4.9.0.0
+-- @since 4.8.1.0
 data CallStack = CallStack { getCallStack :: [([Char], SrcLoc)] }
   -- See Note [Overview of implicit CallStacks]
 
--- | A single location in the source code.
+
+-- Note [Definition of CallStack]
+-- ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+-- Implicit CallStacks are defined very early in base because they are
+-- used by error and undefined. At this point in the dependency graph,
+-- we do not have enough functionality to (conveniently) write a nice
+-- pretty-printer for CallStack. The sensible place to define the
+-- pretty-printer would be GHC.Stack, which is the main access point,
+-- but unfortunately GHC.Stack imports GHC.Exception, which *needs*
+-- the pretty-printer. So the CallStack type and functions are split
+-- between three modules:
+--
+-- 1. GHC.Stack.Types: defines the type and *simple* functions
+-- 2. GHC.Exception: defines the pretty-printer
+-- 3. GHC.Stack: exports everything and acts as the main access point
+
+
+-- | Push a call-site onto the stack.
 --
 -- @since 4.9.0.0
+pushCallStack :: ([Char], SrcLoc) -> CallStack -> CallStack
+pushCallStack callSite (CallStack stk)
+  = CallStack (callSite : stk)
+
+
+-- | A single location in the source code.
+--
+-- @since 4.8.1.0
 data SrcLoc = SrcLoc
   { srcLocPackage   :: [Char]
   , srcLocModule    :: [Char]